chore: moved test code files to types folder #4
