+++ package.json
@@ -52,7 +52,7 @@
     "prettier": "^3.4.2",
     "rimraf": "^6.0.1",
     "ts-patch": "^3.3.0",
-    "typescript": "^5.7.3",
+    "typescript": "~5.7.3",
     "typia": "^7.6.0"
   },
   "repository": {
