+++ test/index.ts
@@ -2,11 +2,24 @@ import fs from 'fs';
 import path from 'path';
 
 async function main() {
-  const files = fs.readdirSync(path.join(__dirname, './'));
-
-  for await (const filename of files) {
+  /**
+   * test code for types
+   */
+  const types = fs.readdirSync(path.join(__dirname, './types'));
+  for await (const filename of types) {
     if (filename.endsWith('.js')) {
-      const filePath = path.join(__dirname, filename);
+      const filePath = path.join(__dirname, 'types', filename);
+      await import(filePath);
+    }
+  }
+
+  /**
+   * test code for functions
+   */
+  const functions = fs.readdirSync(path.join(__dirname, './functions'));
+  for await (const filename of functions) {
+    if (filename.endsWith('.js')) {
+      const filePath = path.join(__dirname, 'functions', filename);
       await import(filePath);
     }
   }
